se reordenaron los archivos de vista genericos en una carpeta para facilitar la lectura y navegacion del proyecto
