+++ src/front/pages/vistas genericas/PoliticaCancelacion.jsx
