+++ src/front/pages/vistas genericas/PoliticaPrivacidad.jsx
