+++ src/front/routes.jsx
@@ -8,17 +8,17 @@ import {
 // home y layout
 import { Layout } from "./pages/Layout";
 import { Home } from "./pages/Home";
-import { Single } from "./pages/Single";
-import { Demo } from "./pages/Demo";
+import { Single } from "./pages/vistas genericas/Single";
+import { Demo } from "./pages/vistas genericas/Demo";
 
 // vistas genericas
-import { QuienesSomos } from "./pages/QuienesSomos";
-import { Contacto } from "./pages/Contacto";
-import { PreguntasFrecuentes } from "./pages/PreguntasFrecuentes";
-import { PoliticaPrivacidad } from "./pages/PoliticaPrivacidad";
-import { TerminosCondiciones } from "./pages/TerminosCondiciones";
-import { PoliticaCancelacion } from "./pages/PoliticaCancelacion";
-import { NotFound } from "./pages/NotFound";
+import { QuienesSomos } from "./pages/vistas genericas/QuienesSomos";
+import { Contacto } from "./pages/vistas genericas/Contacto";
+import { PreguntasFrecuentes } from "./pages/vistas genericas/PreguntasFrecuentes";
+import { PoliticaPrivacidad } from "./pages/vistas genericas/PoliticaPrivacidad";
+import { TerminosCondiciones } from "./pages/vistas genericas/TerminosCondiciones";
+import { PoliticaCancelacion } from "./pages/vistas genericas/PoliticaCancelacion";
+import { NotFound } from "./pages/vistas genericas/NotFound";
 
 //inicio de sesion - registro y gestion de usuarios
 import { Registro } from "./pages/Registro";
@@ -34,11 +34,11 @@ import { Checkout } from "./pages/Checkout";
 
 //vistas admin
 
-import { Customers } from "./pages/AdminPages/Customers"
-import { Dashboard } from "./pages/AdminPages/Dashboard"
-import { Orders } from "./pages/AdminPages/Orders"
-import { Stock } from "./pages/AdminPages/Stock"
-import { Users } from "./pages/AdminPages/Users"
+import { Customers } from "./pages/AdminPages/Customers";
+import { Dashboard } from "./pages/AdminPages/Dashboard";
+import { Orders } from "./pages/AdminPages/Orders";
+import { Stock } from "./pages/AdminPages/Stock";
+import { Users } from "./pages/AdminPages/Users";
 
 export const router = createBrowserRouter(
   createRoutesFromElements(
@@ -55,32 +55,30 @@ export const router = createBrowserRouter(
       <Route path="/single/:theId" element={<Single />} />{" "}
       {/* Dynamic route for single items */}
       <Route path="/demo" element={<Demo />} />
-
-      {/* vistas genericas */} 
+      {/* vistas genericas */}
       <Route path="/quienes-somos" element={<QuienesSomos />} />
       <Route path="/preguntas-frecuentes" element={<PreguntasFrecuentes />} />
       <Route path="/politica-de-privacidad" element={<PoliticaPrivacidad />} />
       <Route path="/confirmacion-compra" element={<ConfirmacionCompra />} />
-      <Route path="/politica-de-cancelacion" element={<PoliticaCancelacion />} />
+      <Route
+        path="/politica-de-cancelacion"
+        element={<PoliticaCancelacion />}
+      />
       <Route path="/contactanos" element={<Contacto />} />
       <Route path="/terminos-y-condiciones" element={<TerminosCondiciones />} />
-
       {/* inicio de sesion y manejo de usuario */}
       <Route path="/registro" element={<Registro />} />
       <Route path="/iniciar-sesion" element={<IniciarSesion />} />
       <Route path="/olvido-su-contraseña" element={<OlvidoContraseña />} />
       <Route path="/recuperar-contraseña" element={<RecuperarContraseña />} />
-
       {/* carrito y vista de producto */}
       <Route path="/cart" element={<Cart />} />
       <Route path="/producto/:id" element={<VistaProducto />} />
       <Route path="/checkout" element={<Checkout />} />
       <Route path="/confirmacion" element={<ConfirmacionCompra />} />
-
       {/* vistas admin */}
       <Route path="/admin" />
-
-      <Route path="*" element={ <NotFound/> } />
+      <Route path="*" element={<NotFound />} />
     </Route>
   )
 );
